Refactor OpCodes.
